Move extractors into their own folder
